feat(subsonic): 实现MediaAnnotation API
添加`MediaAnnotationRepository`以支持媒体注释功能，包括：
- `scrobble`: 记录媒体文件播放。
- `setRating`: 设置歌曲、专辑或艺术家的评分。
- `star`: 为歌曲、专辑或艺术家添加星标。
- `unstar`: 为歌曲、专辑或艺术家移除星标。

同时，修复`setRating`接口中`rating`参数的类型，从`Long`更正为`Int`，以匹配API规范。

+++ lib_open_subsonic_api/src/main/java/com/che2n3jigw/android/libs/opensubsonicapi/service/MediaAnnotationService.kt
@@ -53,7 +53,7 @@ interface MediaAnnotationService {
     @GET("/rest/setRating")
     suspend fun setRating(
         @Query("id") id: String,
-        @Query("rating") rating: Long
+        @Query("rating") rating: Int
     ): BaseResponse<BaseSubsonicResponse>
 
     /**
